+++ src/Magnifier.pro
@@ -33,6 +33,7 @@ SOURCES += \
     INIReader.cpp \
     ini.c \
     MotionDetection.cpp \
+    RieszTransform.cpp \
     main.cpp \
     #
